+++ packages/app/src/components/views/publication/CreateArticleView.tsx
@@ -1,87 +1,31 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 import { Box, FormHelperText, Grid, InputLabel, Stack, TextField } from "@mui/material"
 
-import React, { useEffect, useState } from "react"
-import { INITIAL_ARTICLE_VALUE, usePublicationContext } from "../../../services/publications/contexts"
+import React, { useEffect } from "react"
+import { usePublicationContext } from "../../../services/publications/contexts"
 import { ViewContainer } from "../../commons/ViewContainer"
 import CreateArticlePage from "../../layout/CreateArticlePage"
-import { Article } from "../../../models/publication"
-import { useNavigate, useParams } from "react-router-dom"
-import { useWeb3React } from "@web3-react/core"
-import usePoster from "../../../services/poster/hooks/usePoster"
-import usePublication from "../../../services/publications/hooks/usePublication"
-
 import { ArticleContentSection } from "./components/ArticleContentSection"
-
-import { RICH_TEXT_ELEMENTS } from "../../commons/RichText"
 import { useIpfs } from "../../../hooks/useIpfs"
-import { Block } from "../../commons/EditableItemBlock"
-
-import useLocalStorage from "../../../hooks/useLocalStorage"
-import { Pinning } from "../../../models/pinning"
 import { palette } from "../../../theme"
 import { checkTag } from "../../../utils/string-handler"
-import useMarkdown from "../../../hooks/useMarkdown"
-import useArticles from "../../../services/publications/hooks/useArticles"
 
 interface CreateArticleViewProps {
   type: "new" | "edit"
 }
 
 export const CreateArticleView: React.FC<CreateArticleViewProps> = ({ type }) => {
-  const navigate = useNavigate()
-  const { publicationSlug } = useParams<{ publicationSlug: string }>()
-  const { account } = useWeb3React()
-
   const ipfs = useIpfs()
   const {
     publication,
     draftArticle,
     saveDraftArticle,
-    articleContent,
-    executeArticleTransaction,
-    setExecuteArticleTransaction,
-    draftArticleThumbnail,
-    setLoading,
     article,
     setArticleContent,
     isEditing,
-    setIsIndexing,
+    articleContentError,
+    articleTitleError,
   } = usePublicationContext()
-  const { transactionCompleted } = usePublication(publicationSlug || "")
-  const [pinning] = useLocalStorage<Pinning | undefined>("pinning", undefined)
-  const { createArticle, updateArticle } = usePoster()
-  const {
-    indexing: createArticleIndexing,
-    setExecutePollInterval: createPoll,
-    transactionCompleted: newArticleTransaction,
-    newArticleId,
-  } = useArticles()
-  const {
-    indexing: updateArticleIndexing,
-    setExecutePollInterval: updatePoll,
-    transactionCompleted: updateTransaction,
-    newArticleId: updateArticleId,
-    setArticleId,
-    setCurrentTimestamp,
-  } = useArticles()
-  const { convertToHtml } = useMarkdown()
-  const [titleError, setTitleError] = useState<boolean>(false)
-  const [articleContentError, setArticleContentError] = useState<boolean>(false)
-
-  /*
-   * Component will unmount
-   */
-  useEffect(() => {
-    return () => {
-      setTitleError(false)
-      setArticleContentError(false)
-    }
-  }, [])
-
-  useEffect(() => {
-    setIsIndexing(createArticleIndexing || updateArticleIndexing)
-  }, [createArticleIndexing, updateArticleIndexing])
 
   useEffect(() => {
     if (article && !isEditing) {
@@ -108,162 +52,8 @@ export const CreateArticleView: React.FC<CreateArticleViewProps> = ({ type }) =>
     }
   }, [article])
 
-  /**
-   * Execute transaction
-   */
-  useEffect(() => {
-    const execute = async () => {
-      if (!draftArticle && executeArticleTransaction) {
-        setTitleError(true)
-        setArticleContentError(true)
-        setExecuteArticleTransaction(false)
-        return
-      }
-      if (articleContent && executeArticleTransaction && draftArticle) {
-        await prepareTransaction(articleContent)
-      }
-    }
-    execute()
-  }, [executeArticleTransaction, articleContent])
-
-  /**
-   * Logic after complete the transaction
-   */
-  useEffect(() => {
-    if (transactionCompleted) {
-      saveDraftArticle(INITIAL_ARTICLE_VALUE)
-      navigate(-1)
-    }
-  }, [navigate, saveDraftArticle, transactionCompleted])
-
-  useEffect(() => {
-    if ((newArticleTransaction || updateTransaction) && publicationSlug) {
-      setLoading(false)
-      navigate(`/${publicationSlug}/${newArticleId || updateArticleId}`)
-    }
-  }, [navigate, newArticleId, newArticleTransaction, publicationSlug, updateArticleId, updateTransaction])
-
-  const prepareTransaction = async (articleContent: Block[]) => {
-    let error = false
-    if (draftArticle?.title === "") {
-      setTitleError(true)
-      error = true
-    }
-    if (
-      articleContent.length === 1 &&
-      articleContent[0].html === "" &&
-      articleContent[0].tag !== RICH_TEXT_ELEMENTS.IMAGE
-    ) {
-      setArticleContentError(true)
-      error = true
-    }
-    if (error) {
-      setExecuteArticleTransaction(false)
-      return
-    }
-    setTitleError(false)
-    setArticleContentError(false)
-    setLoading(true)
-    const blocks: Block[] = []
-    for (const block of articleContent) {
-      if (block.tag === RICH_TEXT_ELEMENTS.IMAGE && block.imageFile) {
-        try {
-          await ipfs.uploadContent(block.imageFile).then(async (img) => {
-            blocks.push({ ...block, imageUrl: img.path })
-          })
-        } catch {
-          setLoading(false)
-        }
-      } else {
-        blocks.push(block)
-      }
-    }
-
-    //Validate if the article content exist
-    if (blocks.length === 1 && blocks[0].html === "" && blocks[0].tag !== RICH_TEXT_ELEMENTS.IMAGE) {
-      setLoading(false)
-      return setArticleContentError(true)
-    }
-
-    const content = await convertToHtml(blocks, false)
-
-    if (draftArticle) {
-      const newArticle = { ...draftArticle, article: content }
-      saveDraftArticle(newArticle)
-      await handleArticleAction(newArticle)
-    }
-    setExecuteArticleTransaction(false)
-  }
-
-  const handleArticleAction = async (article: Article) => {
-    let articleThumbnail = ""
-    let hashArticle
-    const { title, article: draftArticleText, description, tags } = article
-    if (draftArticleThumbnail) {
-      await ipfs.uploadContent(draftArticleThumbnail).then(async (img) => {
-        articleThumbnail = img.path
-      })
-    }
-    if (publication && article && account) {
-      const id = publication.id
-      if (id == null) {
-        throw new Error("Publication id is null")
-      }
-      if (pinning && draftArticleText) {
-        hashArticle = await ipfs.uploadContent(draftArticleText)
-      }
-
-      if (title) {
-        if (type === "new") {
-          return await createArticle(
-            {
-              action: "article/create",
-              publicationId: id,
-              title,
-              article: hashArticle ? hashArticle.path : draftArticleText,
-              description,
-              tags,
-              image: articleThumbnail,
-              authors: [account],
-            },
-            hashArticle ? true : false,
-          ).then((res) => {
-            if (res?.error) {
-              setLoading(false)
-            } else {
-              createPoll(true)
-            }
-          })
-        }
-        if (type === "edit" && article && article.id) {
-          await updateArticle(
-            {
-              action: "article/update",
-              id: article.id,
-              title,
-              article: hashArticle ? hashArticle.path : draftArticleText,
-              description,
-              tags,
-              image: articleThumbnail,
-              authors: [account],
-            },
-            hashArticle ? true : false,
-          ).then((res) => {
-            if (res && res.error) {
-              setLoading(false)
-            } else if (article && article.lastUpdated) {
-              setArticleId(article.id)
-              setCurrentTimestamp(parseInt(article.lastUpdated))
-              updatePoll(true)
-            }
-          })
-        }
-      }
-    }
-  }
-
   return (
-    <CreateArticlePage publication={publication}>
+    <CreateArticlePage publication={publication} type={type}>
       <Box
         onSubmit={(e) => {
           e.preventDefault()
@@ -287,7 +77,7 @@ export const CreateArticleView: React.FC<CreateArticleViewProps> = ({ type }) =>
                   placeholder="Post title"
                 />
 
-                {titleError && (
+                {articleTitleError && (
                   <FormHelperText sx={{ color: palette.secondary[1000], textTransform: "capitalize" }}>
                     Title is required
                   </FormHelperText>
