+++ packages/app/src/services/publications/contexts/publication.context.tsx
@@ -35,7 +35,9 @@ const PublicationProvider = ({ children }: PublicationProviderProps) => {
     undefined,
   )
   const [removePublicationImage, setRemovePublicationImage] = useState<boolean>(false)
-  const [isIndexing, setIsIndexing] = useState<boolean>(false)
+  const [articleTitleError, setArticleTitleError] = useState<boolean>(false)
+  const [articleContentError, setArticleContentError] = useState<boolean>(false)
+
   const getPublicationId = async (publicationSlug: string, provider?: ethers.providers.BaseProvider) => {
     if (slugToPublicationId[publicationSlug]) {
       return slugToPublicationId[publicationSlug]
@@ -101,8 +103,10 @@ const PublicationProvider = ({ children }: PublicationProviderProps) => {
         setExecuteArticleTransaction,
         publicationAvatar,
         removePublicationImage,
-        isIndexing,
-        setIsIndexing,
+        articleTitleError,
+        setArticleTitleError,
+        articleContentError,
+        setArticleContentError,
         setRemovePublicationImage,
         setPublicationAvatar,
         setMarkdownArticle,
