+++ packages/app/src/services/publications/contexts/publication.types.ts
@@ -28,8 +28,10 @@ export type PublicationContextType = {
   setPublicationAvatar: (uri: { publicationId: string; uri: string } | undefined) => void
   removePublicationImage: boolean
   setRemovePublicationImage: (remove: boolean) => void
-  isIndexing: boolean
-  setIsIndexing: (indexing: boolean) => void
+  articleContentError: boolean
+  setArticleContentError: (error: boolean) => void
+  articleTitleError: boolean
+  setArticleTitleError: (remove: boolean) => void
   getPublicationId: (publicationSlug: string, provider?: ethers.providers.BaseProvider) => Promise<string | undefined>
   setMarkdownArticle: (markdown: string | undefined) => void
   saveIsEditing: (isEditing: boolean) => void
